Refactor Audio params constants
